+++ lib/features/exercise/presentation/pages/exercise/exercise_form_page.dart
@@ -5,7 +5,10 @@ import '../../../../../core/util/spacing.dart';
 import '../../../domain/entities/exercise.dart';
 import '../../../domain/entities/value_objects/exercise_name.dart';
 import '../../../domain/entities/value_objects/incrementation.dart';
+import '../../../domain/entities/value_objects/month.dart';
 import '../../../domain/entities/value_objects/one_rm.dart';
+import '../../../domain/entities/value_objects/week.dart';
+import '../../../domain/entities/week_names.dart';
 import '../../bloc/exercise_bloc.dart';
 import '../../widgets/centered_loading.dart';
 import '../../widgets/inputs/exercise_name_input.dart';
@@ -76,6 +79,8 @@ class _ExerciseFormState extends State<_ExerciseForm> {
         oneRm: OneRm.parse(_oneRmController.value.text),
         name: ExerciseName(_exerciseNameController.value.text),
         incrementation: Incrementation.parse(_incrementationController.value.text),
+        month: Month(1),
+        week: Week(Weeks.accumulation),
       );
 
   @override